requirements.txt
discord.py>=1.6.0,<=2.0.0
Babel
